Google PubSub sender - load Google PubSub beans only when it is possible to authenticate
